Moved filedrop into its own manager and now I have better control over the actual notch, i'm using a nice flow
