+++ ComfyNotch/ComfyNotchView.swift
@@ -32,9 +32,6 @@ class PanelAnimationState: ObservableObject {
     
     @Published var utilsSelectedTab : UtilsTab = .clipboard
     
-    @Published var droppedFiles: [URL] = []
-    @Published var droppedFile: URL?
-    
     let hoverHandler = HoverHandler()
     
     init() {
@@ -46,11 +43,12 @@ struct ComfyNotchView: View {
     @EnvironmentObject var widgetStore: CompactWidgetsStore
     @EnvironmentObject var bigWidgetStore: ExpandedWidgetsStore
     
+    @StateObject private var fileDropManager = FileDropManager()
+    
     @ObservedObject var animationState = PanelAnimationState.shared
     @ObservedObject var settings = SettingsModel.shared
     
     @Binding private var isDroppingFiles: Bool
-    @Binding private var droppedFiles: [URL]
     
     /// Testing:
     @State private var dragProgress: CGFloat = 0
@@ -64,13 +62,8 @@ struct ComfyNotchView: View {
             get: { panelAnimationState.isDroppingFiles },
             set: { panelAnimationState.isDroppingFiles = $0 }
         )
-        let droppedFilesBinding = Binding<[URL]> (
-            get: { panelAnimationState.droppedFiles },
-            set: { panelAnimationState.droppedFiles = $0 }
-        )
         
         _isDroppingFiles = isDroppingFilesBinding
-        _droppedFiles = droppedFilesBinding
     }
     
     var body: some View {
@@ -81,7 +74,7 @@ struct ComfyNotchView: View {
                 .contentShape(Rectangle())
                 .padding(-100) // expands hit area
                 .onDrop(of: [UTType.fileURL.identifier, UTType.image.identifier], isTargeted: $isDroppingFiles) { providers in
-                    handleDrop(providers: providers)
+                    fileDropManager.handleDrop(providers: providers)
                 }
             
             VStack(alignment: .leading,spacing: 0) {
@@ -92,7 +85,7 @@ struct ComfyNotchView: View {
                 /// see QuickAccessWidget.swift file to see how it works
                 switch animationState.currentPanelState {
                 case .home:         HomeNotchView().environmentObject(bigWidgetStore)
-                case .file_tray:    FileTrayView()
+                case .file_tray:    FileTrayView().environmentObject(fileDropManager)
                 case .messages:     MessagesView()
                 case .utils:        UtilsView()
                 case .popInPresentation: PopInPresenter()
@@ -211,91 +204,5 @@ struct ComfyNotchView: View {
                 }
             }
         }
-    }
-    
-    func handleDrop(providers: [NSItemProvider]) -> Bool {
-        
-        let fm = FileManager.default
-        let sessionDir = fm.temporaryDirectory.appendingPathComponent("FileDropperSession-\(UUID().uuidString)", isDirectory: true)
-        try? fm.createDirectory(at: sessionDir, withIntermediateDirectories: true)
-        
-        for provider in providers {
-            
-            /// ---------- Finder files ----------
-            if provider.hasItemConformingToTypeIdentifier(UTType.fileURL.identifier) {
-                
-                provider.loadInPlaceFileRepresentation(forTypeIdentifier: UTType.fileURL.identifier) {
-                    url, inPlace, _ in
-                    guard let srcURL = url else { return }
-                    
-                    processFile(at: srcURL,
-                                copyIfNeeded: !inPlace,
-                                sessionDir: sessionDir)
-                    
-                }
-                
-                // ---------- Images / screenshots ----------
-            } else if provider.canLoadObject(ofClass: NSImage.self) {
-                _ = provider.loadObject(ofClass: NSImage.self) { object, _ in
-                    guard let img = object as? NSImage,
-                          let tiff = img.tiffRepresentation,
-                          let rep  = NSBitmapImageRep(data: tiff),
-                          let png  = rep.representation(using: .png, properties: [:])
-                    else { return }
-                    
-                    let tmpURL = sessionDir.appendingPathComponent(
-                        "DroppedImage-\(UUID()).png")
-                    
-                    Task.detached(priority: .utility) {
-                        try? png.write(to: tmpURL)   // fast, one write
-                        await processFile(at: tmpURL,
-                                          copyIfNeeded: false,
-                                          sessionDir: sessionDir)
-                    }
-                }
-            }
-            
-            // ---------- Promised files ----------
-            else if provider.registeredTypeIdentifiers.contains("com.apple.filepromise") {
-                provider.loadDataRepresentation(forTypeIdentifier: "com.apple.filepromise") { _, error in
-                    if let error = error {
-                        debugLog("❌ Failed to receive file promise: \(error)")
-                        return
-                    }
-                    debugLog("ℹ️ File promise received — but not handled in this version")
-                }
-            }
-        }
-        
-        return true
-    }
-    
-    private func processFile(at url: URL,
-                             copyIfNeeded: Bool,
-                             sessionDir: URL) {
-        Task.detached(priority: .utility) {
-            //            guard let (size, hash) = DroppedFileTracker.shared.quickHash(url: url),
-            //                  DroppedFileTracker.shared.isNewFile(size: size, hash: hash) else {
-            //                debugLog("Duplicate File Detected: \(url)")
-            //                return
-            //            }
-            
-            let settings = SettingsModel.shared
-            let saveFolder = settings.fileTrayDefaultFolder
-            
-            try? FileManager.default.createDirectory(at: saveFolder, withIntermediateDirectories: true)
-            let destURL = saveFolder.appendingPathComponent(url.lastPathComponent)
-            let sourceURL = copyIfNeeded ? url : url // ← future-proof
-            try? FileManager.default.copyItem(at: sourceURL, to: destURL)
-            
-            //            DroppedFileTracker.shared.registerFile(size: size,
-            //                                                   hash: hash,
-            //                                                   url: destURL)
-            
-            // 3. Tell SwiftUI
-            await MainActor.run {
-                PanelAnimationState.shared.droppedFile = destURL
-            }
-        }
-    }
+    }    
 }
